+++ Example/PaymentSheet Example/PaymentSheet Example/ExampleWalletButtonsView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 struct ExampleWalletButtonsContainerView: View {
     @State private var email: String = ""
-    @State private var shopId: String = "shop_id_123"
+    @State private var shopId: String = ""
     @State private var linkInlineVerificationEnabled: Bool = PaymentSheet.LinkFeatureFlags.enableLinkInlineVerification
     @State private var useSPTTestBackend: Bool = false
     @State private var appearance: PaymentSheet.Appearance = PaymentSheet.Appearance()
@@ -517,7 +517,7 @@ class ExampleWalletButtonsModel: ObservableObject {
                                                  handlers: handlers)
     }
     func isValidShippingLocation(_ address: PaymentSheet.ShopPayConfiguration.PartialAddress) -> Bool {
-        return true
+        return address.postalCode != "91911"
     }
 }